Commit de Ronan Le Boulch : - Correction d'une erreur de sqrt sous windows
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@1719 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/dct_interface.cpp
@@ -190,7 +190,7 @@ MP_DCTW_Interface_c::MP_DCTW_Interface_c( const unsigned long int setDctSize )
 
   /* Create plans */
   p = fftw_plan_r2r_1d( (int)(dctSize), inPrepared, out, FFTW_REDFT11, FFTW_MEASURE );
-  scale = 1/sqrt(dctSize);
+  scale = 1/sqrt((double)dctSize);
 }
 
 /**************/
